Module cimdea::ipums_data_model

Model the IPUMS datasets (sometimes called ‘samples’.) This module models the record types, weighting and record type relationship for IPUMS data, not the samples themselves.

Every dataset in a collection will have these same characteristics: It will be a collection of record types and each record type has some information about it and its relationships to the other record types. Every collection has a single hierarchy of record types. A record type on a particular data product may have a default weight variable – or it may not.

Structs

RecordHierarchy

RecordHierarchyMember

RecordType

RecordWeight