Struct cimdea::request::AbacusRequest

pub struct AbacusRequest {
    pub product: String,
    pub request_variables: Vec<RequestVariable>,
    pub subpopulation: Vec<RequestVariable>,
    pub request_samples: Vec<RequestSample>,
    pub unit_rectype: RecordType,
    pub output_format: OutputFormat,
    pub use_general_variables: bool,
    pub data_root: Option<String>,
}

The Abacus Request type contains variables to tabulate, variables used for conditions and datasets.

Fields
product: Stringrequest_variables: Vec<RequestVariable>subpopulation: Vec<RequestVariable>request_samples: Vec<RequestSample>unit_rectype: RecordTypeoutput_format: OutputFormatuse_general_variables: booldata_root: Option<String>
Implementations

impl AbacusRequest

pub fn try_from_json(input: &str) -> Result<(Context, Self), MdError>

Accepts a single JSON with keys for Request, Subpop and any other arguments. /// { โproductโ: โusaโ, โdata_rootโ : โ/pkg/ipums/usa/output_data/currentโ, โrequest_variablesโ: [โฆ], request_samples: [โฆ], โsubpopโ : [ {โฆ}, {โฆ}], โuoaโ : โPโ}