Trait cimdea::request::DataRequest

pub trait DataRequest {
    // Required methods
    fn get_request_variables(&self) -> Vec<RequestVariable>;
    fn get_request_samples(&self) -> Vec<RequestSample>;
    fn get_conditions(&self) -> Option<Vec<Condition>>;
    fn serialize_to_ipums_json(&self) -> String;
    fn deserialize_from_ipums_json(
        ctx: &Context,
        request_type: RequestType,
        json_request: &str,
    ) -> Result<Self, MdError>
       where Self: Sized;
    fn from_names(
        product_name: &str,
        requested_datasets: &[&str],
        requested_variables: &[&str],
        unit_of_analysis: Option<String>,
        optional_product_root: Option<String>,
        optional_data_root: Option<String>,
    ) -> Result<(Context, Self), MdError>
       where Self: Sized;
    fn print_codebook(&self) -> String;
    fn print_stata(&self) -> String;
    fn case_select_logic(&self) -> CaseSelectLogic;
    fn case_select_unit(&self) -> CaseSelectUnit;
}

Every data request should serialize, deserialize, and produce SQL queries for what it’s requesting.

Required Methods

fn get_request_variables(&self) -> Vec<RequestVariable>

fn get_request_samples(&self) -> Vec<RequestSample>

fn get_conditions(&self) -> Option<Vec<Condition>>

fn serialize_to_ipums_json(&self) -> String

Convert to the Tractor / generic IPUMS representation

fn deserialize_from_ipums_json( ctx: &Context, request_type: RequestType, json_request: &str, ) -> Result<Self, MdError>
where Self: Sized,

Convert from the Tractor / generic JSON representation.

fn from_names( product_name: &str, requested_datasets: &[&str], requested_variables: &[&str], unit_of_analysis: Option<String>, optional_product_root: Option<String>, optional_data_root: Option<String>, ) -> Result<(Context, Self), MdError>
where Self: Sized,

Build request from a basic set of variable and dataset names and data locations. Get back a SimpleRequest and a Context needed to execute the request or future requests.

fn print_codebook(&self) -> String

Print a human readable codebook

fn print_stata(&self) -> String

Print a machine readable Stata codebook

fn case_select_logic(&self) -> CaseSelectLogic

fn case_select_unit(&self) -> CaseSelectUnit