Crate cimdea

Modules

conventions

This module provides structs and methods for loading metadata and storing information about a IPUMS data collection based on conventions and minimal configuration. Every collection has a set of data record types and a hierarchy those records belong to: For instance, person records belong to household records.

defaults

The Household - Person record structure is the default for much IPUMS data. Here we have some functions to support setting up such a default structure without needing any external configuration.

ipums_data_model

Every dataset in a collection will have these same characteristics: It will be a collection of record types and each record type has some information about it and its relationships to the other record types. Every collection has a single hierarchy of record types. A record type on a particular data product may have a default weight variable – or it may not.

ipums_metadata_model

The “metadata” models serve to assist working with IPUMS data. The entities here match the full IPUMS metadata in terms of their relationships to one another and their description of the data. However they don’t include (1) all fields / pieces of info from the full IPUMS metadata; and (2) do not contain all metadata models – only those essential for understanding the data files on a technical level. For instance there are no enumeration text or citations metadata modeled here.

layout

Supports reading the “layout” files from IPUMS microdata. These files describe fixed-width formatted IPUMS datasets. While they were developed to act as minimal record layouts for fixed-width they can be useful for getting basic metadata for the dataset.

query_gen

This is an exploratory module to try out ideas for aggregating IPUMS data with generic SQL. Instead of the DB specific query builders and parameter binding, see if we can do it in a generic way TODO For Duck DB we need the table name if it’s parquet to looke like 'table_name.parquet' and it needs to be a valid file, or if the parquet is multiple files it would look like table_name/*.parquet'. /

request

Requests describe requested IPUMS data either as an extract – multiple records – or an aggregation of data from those records. We use names of IPUMS “variables” and “datasets” in the requests along with other details.