Crate cimdea

CIMDEA == Convenientt IPUMS Microdata Extraction and Aggregation

Modules

conventions

This module provides structs and methods for loading metadata and storing information about a IPUMS data collection based on IPUMS conventions and minimal configuration.

defaults

The Household - Person record structure is the default for much IPUMS data. Here we have some functions to support setting up such a default structure without needing any external configuration. Everything modeled here could originate from a run-time configuration process instead.

ipums_data_model

Model the IPUMS datasets (sometimes called ‘samples’.) This module models the record types, weighting and record type relationship for IPUMS data, not the samples themselves.

ipums_metadata_model

The “metadata” models serve to assist working with IPUMS data. The entities here match the full IPUMS metadata in terms of their relationships to one another and their description of the data.

layout

Supports reading the “layout” metadata files from IPUMS microdata.

query_gen

Generate queries from a DataRequest . Currently supports cross-tab style queries.

request

Requests describe requested IPUMS data either as an extract – multiple records – or an aggregation of data from those records. We use names of IPUMS “variables” and “datasets” to make the requests along with other details.

tabulate

The high level module for executing and formatting tabulations.